# web/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/blog/blogs" : {
      "get" : {
        "summary" : "获取博客列表 - RESTful GET（支持分页和条件查询）",
        "description" : "获取博客列表 - RESTful GET（支持分页和条件查询）",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : 1
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : 10
          }
        }, {
          "name" : "cateId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "title",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "state",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "博客分页列表",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "records" : {
                          "type" : "array",
                          "description" : "",
                          "items" : {
                            "type" : "object",
                            "properties" : {
                              "id" : {
                                "type" : "string",
                                "description" : ""
                              },
                              "title" : {
                                "type" : "string",
                                "description" : ""
                              },
                              "cateId" : {
                                "type" : "string",
                                "description" : ""
                              },
                              "cateName" : {
                                "type" : "string",
                                "description" : ""
                              },
                              "coverImg" : {
                                "type" : "string",
                                "description" : ""
                              },
                              "content" : {
                                "type" : "string",
                                "description" : ""
                              },
                              "state" : {
                                "type" : "integer",
                                "description" : ""
                              },
                              "createTime" : {
                                "type" : "string",
                                "description" : ""
                              },
                              "updateTime" : {
                                "type" : "string",
                                "description" : ""
                              }
                            },
                            "description" : "com.gaoyifeng.aioserver.api.dto.blog.response.BlogGetListResponseDto"
                          }
                        },
                        "total" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "pages" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "current" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "size" : {
                          "type" : "integer",
                          "description" : ""
                        }
                      },
                      "description" : ""
                    }
                  },
                  "description" : "博客分页列表"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "创建博客 - RESTful POST",
        "description" : "创建博客 - RESTful POST",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "title" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "cateId" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "coverImg" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "content" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "state" : {
                    "type" : "integer",
                    "description" : ""
                  }
                },
                "description" : "添加博客请求DTO"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "添加结果",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  },
                  "description" : "添加结果"
                }
              }
            }
          }
        }
      }
    },
    "/blog/blogs/{id}" : {
      "put" : {
        "summary" : "编辑博客 - RESTful PUT",
        "description" : "编辑博客 - RESTful PUT",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "博客ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "title" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "cateId" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "coverImg" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "content" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "state" : {
                    "type" : "integer",
                    "description" : ""
                  }
                },
                "description" : "编辑博客请求DTO"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "编辑结果",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  },
                  "description" : "编辑结果"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "删除博客 - RESTful DELETE (支持单个和批量删除)",
        "description" : "删除博客 - RESTful DELETE (支持单个和批量删除)",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "博客ID (单个删除)",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "ids" : {
                    "type" : "array",
                    "description" : "",
                    "items" : {
                      "type" : "string"
                    }
                  }
                },
                "description" : "批量删除请求DTO (可选，用于批量删除)"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "删除结果",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  },
                  "description" : "删除结果"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}